requirements.txt
djangorestframework==3.14.0
drf-yasg==1.21.7
openai==0.28.0
pandas==2.1.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
tiktoken==0.5.1
